+++ packages/package-json-lint-config-terra/package-json-lint.config.js
@@ -1,7 +1,7 @@
 module.exports = {
   rules: {
-    'require-no-hard-coded-dependency-versions': 'error',
-    'require-no-terra-base-peer-dependency-versions': 'warn',
-    'require-theme-context-versions': 'warn',
+    'require-no-hard-coded-dependency-versions': { severityType: 'error' },
+    'require-no-terra-base-peer-dependency-versions': { severityType: 'warn' },
+    'require-theme-context-versions': { severityType: 'warn' },
   },
 };